pyproject.toml
[tool.poetry]
name = "python-google-gemini"
version = "0.1.0"
description = ""
authors = ["alextanhongpin <alextan220990@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
jupyterlab = "^4.0.9"
google-generativeai = "^0.3.1"
pillow = "^10.1.0"
llama-index = "^0.9.15.post2"
ipywidgets = "^8.1.1"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
